+++ tests/test_nl/forms.py
@@ -1,6 +1,6 @@
 from django.forms import ModelForm
 
-from .models import NLPlace
+from .models import NLCar, NLPlace
 
 
 class NLPlaceForm(ModelForm):
@@ -8,3 +8,10 @@ class NLPlaceForm(ModelForm):
     class Meta:
         model = NLPlace
         fields = ('zipcode', 'province', 'bsn')
+
+
+class NLCarForm(ModelForm):
+
+    class Meta:
+        model = NLCar
+        fields = ('license_plate', )